% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tidy_FPQR.R
\name{build_tidy_FPQR}
\alias{build_tidy_FPQR}
\title{Build Tidy FPQR Wide-Format Data Frame}
\usage{
build_tidy_FPQR(
  d_forms,
  d_tre,
  d_vis,
  d_pat,
  no_kela = T,
  no_kids = T,
  no_interrupted_therapies = T,
  no_repeated_FPQR_treatments = T,
  no_internal_therapies = T
)
}
\arguments{
\item{d_forms}{The BCB forms data table}

\item{d_tre}{The BCB treatments data table}

\item{d_vis}{The BCB visits data table}

\item{d_pat}{The BCB patients data table}

\item{no_kela}{Exclude SSI (Kela) therapies. Logical value. The default.}

\item{no_kids}{Exclude non-adults. Logical value. The default.}

\item{no_interrupted_therapies}{Exclude interrupted therapies. Set FALSE only if you know what you are doing.}

\item{no_repeated_FPQR_treatments}{Exclude reruns. Set FALSE only if you know what you are doing.}

\item{no_internal_therapies}{Exclude HUS internal therapies.}
}
\value{
A list of data frames
}
\description{
Builds a tidy FPQR data. This an early-stage function:
if expecting multiple roles for same patients, the function may have
inadvertent consequences due to ID-based removal of patients in a role.
}
