% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataToBLFU_wvis.R
\name{dataToBLFU_wvis}
\alias{dataToBLFU_wvis}
\title{FPQR item data to baseline-follow-up, or pre-post, format}
\usage{
dataToBLFU_wvis(
  dorig,
  items,
  dvis,
  septh = 30,
  interrupt_ids = NULL,
  use_orig_lab = FALSE
)
}
\arguments{
\item{dorig}{Wide-format FPQR data without pre-post separation}

\item{items}{A string vector of item variable names}

\item{dvis}{FPQR visits data table}

\item{septh}{Separation threshold in days needed to assiqn pre vs. post statuses. Defaults to 30.}

\item{interrupt_ids}{Post-values of optional interrupted therapy IDs are forced to missing values. Defaults to NULL.}

\item{use_orig_lab}{Logical. Use original item labels instead of generic BL and FU. Defaults to FALSE.}
}
\value{
A data frame.
}
\description{
FPQR item data to baseline-follow-up, or pre-post, format
}
\examples{
# Read SOFAS scores from database d to pre-post format
library(tidyr)
library(dplyr)
dsof <- dsof_example \%>\%
filter(template_code == "sofas") \%>\%
select(patient_id, date_created, template_code, question_code, number_answer,
 treatment_id, visit_id) \%>\%
group_by(question_code) \%>\%
mutate(row = row_number()) \%>\% # to define unique rows for pivoting
ungroup() \%>\%
pivot_wider(names_from = question_code, values_from = number_answer) \%>\%
select(-row)

# SOFAS, time, and ID variables
dsof <- subset(dsof, select = c(patient_id, sofas_asteikolla, vapaa_aika,
                                tyo_tai_opiskelu,perhe_elama_ja_ihmissuhteet,
                                itsesta_huolehtiminen,date_created,
                                treatment_id, visit_id))
# Rename from Finnish to English
names(dsof) <- c("patient_id","sofas","freetime","work","family","self",
"date_created","treatment_id","visit_id")

# Finally, call dataToBLFU_wvis
dsof <- dataToBLFU_wvis(dsof, items = names(dsof)[c(2:6)], dvis_example,
interrupt_ids = NULL, use_orig_lab = TRUE)
}
