% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_dependence.R
\name{direction_dependence}
\alias{direction_dependence}
\title{Pairwise Direction of Dependence Inference}
\usage{
direction_dependence(x, y, approximation = "mxnt")
}
\arguments{
\item{x}{A numeric vector (one variable)}

\item{y}{A numeric vector (the other variable)}

\item{approximation}{}
}
\value{
A positive scalar value when x causes y and negative when
other way around. Near-zero values for no direction (use bootstrapping).
}
\description{
Computes the direction of dependence between two
continuous-valued variables \code{x} and \code{y}.
}
