% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantaneous_effect_confounding_test.R
\name{instantaneous_effect_confounding_test}
\alias{instantaneous_effect_confounding_test}
\title{HSIC-based Test for Confounding}
\usage{
instantaneous_effect_confounding_test(X1, X0, dontprint = FALSE)
}
\arguments{
\item{X1}{A matrix with two time 1 variables as columns}

\item{X0}{A matrix with two time 0 variables as columns}

\item{dontprint}{Set to TRUE if you don't want a print out}
}
\value{
Results of the test
}
\description{
HSIC-based test for confounding in the
estimated instantaneous effects.
}
