% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgv_contrast.R
\name{kgv_contrast}
\alias{kgv_contrast}
\title{Kernel Generalized Variance Constrast (KGV) Function}
\usage{
kgv_contrast(
  x,
  sigma = ifelse(nrow(x) <= 1000, 2/100, 2/1000),
  eta = 1e-04,
  kappaval = 0.01
)
}
\arguments{
\item{x}{A data matrix}

\item{sigma}{A default value}

\item{eta}{A default value}

\item{kappaval}{A default value}
}
\value{
A scalar value for the KGV contrast of the data.
}
\description{
This helper function computes the KGV contrast for
kernel-based causal direction dependence estimation in the
direction_dependence and lagged_direction_dependence functions.
}
