% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_direction_dependence.R
\name{lagged_direction_dependence}
\alias{lagged_direction_dependence}
\title{Pairwise Direction Dependence Measures with Lag Included}
\usage{
lagged_direction_dependence(X1, X0, approximation = "mxnt")
}
\arguments{
\item{X1}{A matrix with two time 1 variables as columns}

\item{X0}{A matrix with two time 0 variables as columns}

\item{approximation}{This defaults to the “mxnt” estimator
but may be supplemented with an optional argument
\code{approximation = "mxnt"} (alternatives are mxnt, kgv, tanh, skew,
rskew, hsic; see the original paper for details).}
}
\value{
A list of results
}
\description{
Pairwise Direction Dependence Measures with Lag Included
}
