% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_testdata_insomnia_iCBT.R
\name{make_testdata_insomnia_iCBT}
\alias{make_testdata_insomnia_iCBT}
\title{Make Test Data for Insomnia-iCBT Build Function}
\usage{
make_testdata_insomnia_iCBT(n = 50)
}
\arguments{
\item{n}{Number of patients. Defaults to 50.}
}
\value{
A list of data frames
}
\description{
Make Test Data for Insomnia-iCBT Build Function
}
